Move types to types.ts file
