+++ widgets/common/src/ExecuteWorkflowInputs.tsx
@@ -1,32 +1,11 @@
-import type { ComponentProps, FunctionComponent } from 'react';
-
-export type OnChange = (event: Event, field: typeof Stage.Basic.UnsafelyTypedFormField) => void;
-export type OnDropdownChange = ComponentProps<typeof Stage.Basic.Dropdown>['onChange'];
-export type OnCheckboxChange = ComponentProps<typeof Stage.Basic.Checkbox>['onChange'];
-
-type BaseWorkflowInputs = Record<
-    string,
-    {
-        type?: string;
-        default?: string;
-        constraints?: {
-            pattern: string;
-        }[];
-    }
->;
-
-type UserWorkflowInputsState = Record<
-    string,
-    | string
-    | number
-    | boolean
-    | string[]
-    | number[]
-    | boolean[]
-    | Record<string, string | number | boolean>
-    | null
-    | undefined
->;
+import type { FunctionComponent } from 'react';
+import type {
+    OnChange,
+    OnDropdownChange,
+    OnCheckboxChange,
+    BaseWorkflowInputs,
+    UserWorkflowInputsState
+} from './types';
 
 const t = Stage.Utils.getT('widgets.common.deployments.execute');
